fabric: Add priority fields to msg structures
A socket may support multiple priority queues underneath.
Expose a priority field in the msg structures, so that
the user can select an appropriate queue.  This allows
the fabric socket to abstract multiple hardware queues,
support different SLs on a single UD QP, etc.

This is a more generic version of 'urgent' data.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>