fabric: Add support for direct access to providers
Provide a mechanism for applications to go directly to the
providers without going through the framework or function
vector tables.

Add a configure flag --enable-direct=<provider> to enable direct
calls to a provider that bypass the default static inlines.
    
The provider that was enabled in direct mode installs the files:
    
fi_direct.h
fi_direct_domain.h
fi_direct_endpoint.h
fi_direct_tagged.h
fi_direct_rdma.h
...
    
The user of libfabric compiles with -DFABRIC_DIRECT in order to
use the inline version of the calls. Since the inlined code is
non-portable, it is better to make this explicit to the user.
    
The libfabric header files fabric.h, fi_domain.h, etc. include
their fi_direct_*.h bretheren when being used with -DFABRIC_DIRECT.
The direct files cannot override/redefine any of the basic
libfabric datastructures, and that is made explicit in the placement
of the #ifdefs. The direct includes can only override the static inlines.
    
One of the benefits of this approach is that no #if<PROVIDER> name
is included in the fabric header files.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>