fabric: Replace ERRINLINE with error structures
In order to use ERRINLINE, we need to define a structure
that's the union of fi_ec_err_entry and other fi_ec_XXX_entry
structures.  This is needed in order to report errors along
with successful events.

However, in defining that structure, we can simplify
the interface by adding a new fi_ec_format enum value that
corresponds directly to the new structure.  This eliminates
the need to keep FI_ERRINLINE.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>