fabric: Add fabric domain and passive endpoints
Define a new domain class, the fabric domain.  The fabric
class becomes the top level object presented to applications.
The resource domain is now created under the fabric class.
The fabric class is also associated with passive endpoints
and CM event queues (for listening endpoints).

The endpoint class is separated into active and passive.
Passive endpoints are associated with the fabric class (at
least for now).  All active endpoints are allocated from
an 'access resource domain' (i.e. NIC).  This removes the
need to bind the endpoint to the domain later.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>