fabric: Add enable() call to endpoint
Introduce an enable call with an active endpoint.  Enable
indicates that the endpoint has been configured with all
necessary data and that all necessary HW/SW resources should
be allocated to the endpoint and it be able to support
sending and receiving data.

Without enable, the provider may have a difficult time
determining when the user is done configuring the endpoint.
An endpoint is automatically enabled if connect or accept
is called on the endpoint.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>