fabric/man: Clarify use of olen in fi_eq_tagged_entry
We need to clarify if a receive that overflows the posted
receive buffer completes as successful or not.  We mark overflow
as an error case, on both the send and receive side, and report
the completion as such.  This moves the olen field from
fi_eq_tagged_entry to fi_eq_err_entry, which allows us to
capture overflow from non-tagged transfer types (assuming
that the underlying protocol doesn't just treat overflow
as a fatal error).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>