fabric: Only report error completions through fi_eq_readerr
The EQ interface supports the notion of what was referred to
as 'ERRINLINE', which allowed errors to be reported to the
user through the fi_eq_read() call.  This complicates the
interface for the user (based on feedback from people trying
to understand how the EQ interfaces worked).  Simplify the
usage model by moving all errors from the readerr() method.

We keep the optimized model of reporting only successful
completions through fi_eq_read(), which avoids the app needing
to check both the return code and a request status value to
determine if an operation actually succeeded.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>