fabric: Clarify use of fi_ep_sync
The sync operation only needs to complete after all operations
issued on the endpoint or alias endpoint have completed.
Operations posted to another alias or the base endpoint are
not required to complete.

This supports apps that use flow steering to direct traffic, with
each flow associated with a different alias.  When calling sync on
that alias, only operations posted to that alias are of concern.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>