fabric: Rework endpoint capability and operation flags
There's significant confusion over the use of EP flags and
whether a flag is used to configure the EP, or if the flag
applies to data transfer operations.  In many cases, the same
flag is used both ways.

To clarify this, rename the protocol_cap flag as an ep_cap
flag.  Flags that apply to ep_cap indicate the desired
functionality that an endpoint must provide.  Rename the
fi_info::flags as op_flags, to specify the default flags that
an endpoint should use with data transfer operations.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>