fabric: Add mmap address to AV
In order to share an AV across multiple processes, apps
need a way to share the AV output data (returned fi_addr
values).  The fi_addr is a void *, which is only usable
if the AV data is mapped to the same address.

Add a base-address value to the AV attribute, so that the
returned AV pointers may also be shared between the apps.