Add pkg-config support
From the pkg-config website: "pkg-config is a helper tool used when
compiling applications and libraries. It helps you insert the correct
compiler options on the command line." [1]

Providing a pkg-config file is suggested by the systemd developers as a
best practice in their libabc sample library [2], and allows consumers of
libfabric to use pkg-config in their autotools setup to locate and pull
in libfabric.

[1] http://www.freedesktop.org/wiki/Software/pkg-config/
[2] https://git.kernel.org/cgit/linux/kernel/git/kay/libabc.git/plain/README

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>