fabric/man: Document that an fi_info may only be used once
A single struct fi_info can only be used to allocate a single
endpoint.  The limitation is a result of providers being able to
use the data field to pass data from the fi_getinfo call to
the open endpoint call.  This capability is itself an optimization
made available to the provider, in order to avoid duplicate
memory allocations or computations.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>