fabric: Remove op size field in favor of global version
Remove the size fields from op structures.  Instead, have the
providers give a version when registering with libfabric.  This
reduces the size of the op structures and simplifies adding new
functions in the future.  Libfabric will check that the version
is supportable, and flag any registrations where an older
vendor library is mixed with a newer version of libfabric.

For internal providers, this is a non-issue.  For external
providers, it requires re-compiling the provider libraries
whenever a minor release update occurs to libfabric.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>