prov/psm: remove obsolete mr signature.
In earlier implementation the address of the MR structure is used as
the access key. The signature field is used as a safety measure
against invalid keys. In the current implementation the MR structures
are stored in a table that can be searched by the key. There is no
need to check the signature field.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>