man: document fi_getinfo as a thread safe call
Multiple threads should be able to call fi_getinfo
simultaneously without having to serialize their calls.
This is required to support use cases where multiple
threads may want to access the fabric but they are not aware
of the existence of other threads, or do not have any shared
resource to serialize on.