include: fi_inject_atomic functions
The  fi_inject_atomic call is an optimized version of fi_atomic.  The
fi_inject_atomic function behaves as if the FI_INJECT transfer flag ere
set, and FI_EVENT were not.  That is, the data buffer is available for
reuse immediately on returning from from fi_inject_atomic, and no
completion vent will be generated for this atomic.  The completion event
will be uppressed even  if  the endpoint has not been configured with
FI_EVENT.  See the lags discussion below for more details.

The inject calls are similar to those that are already defined for RMA
and MSGQ.