include: rename FI_INJECT_MSG to FI_OP_INJECT etc.
Rename the macros to check the presence of functions
from FI_INJECT_MSG to FI_OP_INJECT to further clarify
that they check the presence of ops.

Also, updated the fi_getinfo manpage to indicate that
EP cap FI_INJECT may not imply that inject functions are
available for each type of operation.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>