fabric: Replace max_tag_value with mem_tag_format
The maximum tag value allows a provider to indicate the largest tag
that they can support.  Rather than reporting this, use a tag format
instead.  The advantage of a tag format allows the provider and
app to exchange information on not only the size of the tag, but
the general usage of it.  If a tag is used to convey fields of data,
rather than simply individual bits, then a provider may be able
to optimize its matching algorithm.  For example, each field can
represent a separate queue, which would allow for parallel searches,
plus reduce the list of waiting tags to check.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>