fabric: Improve EQ interfaces
Separate the EQ interfaces into control and data EQs.  Control
EQs will be used to report control operations - CM and AV events,
memory registration completions, and other registered events.
Data EQs will be used to report completions of data transfer
operations.

Rename the data EQ as a CQ (aligns better with existing APIs).
Simplify the APIs and data structures for EQs and CQs.  CQs
will become domain level objects.  EQs are fabric level objects.

These changes are needed to clarify when a domain EQ versus
a fabric EQ should be used, and to allow the provider to
distinguish between the two.

As part of this update, include a fix for the verbs provider
where completions may be lost when an error occurs.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>