fabric: Modify CM accept/reject calls to take a connreq context
Update the CM event to report a context associated with the connection
request.  This context will be passed into the accept and reject calls.
This simplifies the CM usage model with stronger type checking.  It
also allows us to remove the 'data' field from fi_info, allowing the
fi_info structure to be independent of the actual connection and
endpoint.  The fi_info will still convey the attributes of the endpoint,
there just won't be an implied association between an instance of the
structure and the connection.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>