fabric: Add provider details to fi_prov structure
Rename fi_ops_prov to fi_prov and add the provider name and version
data to the structure.  This will allow the framework to detect
duplicate providers and select the most recent version of a
provider.  This will allow a provider built into the library to
be overridden with one that is loaded dynamically, and vice-versa.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>