fabric: Define scalable endpoints
Scalable endpoints are endpoints that support multiple send and
receive queues (contexts).

Introduce the concept of transmit and receive contexts
(generalizations of send/receive queues).  Allow applications to
request that multiple contexts be associated with a single endpoint.

An endpoint has a single transport address, but the contexts allow
for multiple threads to process requests in parallel without
ordering or synchronization constraints.

This requires additional domain, endpoint, and AV attributes.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>