configury: re-factor to	extract	common code
Pull common code out of the provider .m4 scripts and back into a
central .m4 file.  Now much of the setup and administration stuff for
FI_<provider>_CONFIGURE is done in a central location -- the
_CONFIGURE macros can just be small chunks of code that focus on
checking to see if that provider can be built (e.g., check for headers
and libraries and that's it).