Clean up dl-open features:
Add --with-dlopen=[yes,no] option to compile without loadable provider support.
Selecting no also disables the check for -ldl.

Rename FI_EXTPATH to FI_EXTDIR for consistency with variable names and configure options.

Remove unused includes in fabric.c

Re-include common.c and enosys.c in dl-loadable providers, these are needed.

Address issues from review:

Remove the ./configure-able extdir: this isn't really needed, if one wants to not
scandir a directory on libfabric load do:

1) build with: --enable-dlopen=no
2) set FI_EXTDIR=/dev/null

Not addressed in this patch:

common sources. This is a seperate issue, in the meantime
re-add these so dl-loadable providers are not broken.

More error-checking in configure script for mistakes (ie dlopen=no and providers
built with =dl). I have more ideas for this and will address in a seperate patch.