Add ability to specify message prefix on posted buffers
This is a proposed patch for #127 .  The PREFIX flags need to be part of
ep_cap (possibly renamed to ep_options or split into ep_cap + ep_attr::ep_flags,
TBD as part of larger rework of ep_cap).

The prefix size, msg_prefix_len, needs to be an ep_attr because its
value will depend on the transport protocol, which is in ep_attr.  i.e.
40 for IB, 48 for UDP.

Signed-off-by: Reese Faucette <rfaucett@cisco.com>