fabric: Introduce 'mode' bitfield into fi_info
There are some features/flags/capabilities/not-sure-what-to-call-this
that a provider needs to expose to an app in order for the app to
get the best performance out of the provider. BUT, the app usually
will NOT want this "feature", so would never request it. The
domain capability FI_LOCAL_MR is an example. Apps don't want to
deal with memory registration, but may need to in order to get
decent performance or work with a provider at all (such as a verbs
provider) it may need to.

Add a mode field to fi_info.  Mode is defined such that an app
will request what modes of operation that it can support.  The
provider will clear mode bits that are not required, and keep
set only those that are.  An app must check the resulting mode
bits that are set and react accordingly.

If an app does not set a mode bit that is required by a provider,
that provider will return FI_ENODATA to the fi_getinfo request.

Note that FI_WRITE_COHERENT moves from a domain capability to a
mode bit.  This results in the meaning of FI_WRITE_COHERENT inverting.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>