fabric: Define FI_MSG_PREFIX mode
Introduce a new operation mode by which the application
provides the buffer space needed to implement the underlying
protocol.  This is an optimization for the provider, which
can avoid it needing to use separate protocol buffers.

The advantage is that a single SGE will contain both the
protocol headers and the application data.  This reduces
the number of buffers that HW must access in order to transmit
the data and places the received headers and data together.

The disadvantage to an app is that it must allocate the
buffer space when sending or receive messages.  FI_MSG_PREFIX
does not apply to RMA data.  Applications that do not want
to support this mode can disable this optimization, which
will require that the provider allocate its own protocol
headers.

Patch is derived from work by Reese Faucette <rfaucett@cisco.com>

Signed-off-by: Sean Hefty <sean.hefty@intel.com>