fabric: Merge ep_cap and domain_cap
Replace the separate capability fields with a single field.  This
field applies to general output, which includes the fabric, domain,
endpoint, and transmit/receive contexts.

This simplifies things for the app, which no longer needs to know
if a capability is an endpoint or domain cap.  It also allows a
single capability to apply to multiple objects collectively.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>