fabric/cq: Rename FI_ADDR_UNSPEC to FI_ADDR_NOTAVAIL
"UNSPEC" is used in other places for an app to signal that they
have not selected a specific value.  In the case of cq readfrom,
we're really wanting to return that no address is available, not
that is is unspecified.

FI_ADDR_UNSPEC will be added back in a later patch, but with
a meaning that aligns with other use cases for 'UNSPEC'.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>