fabric/av: Replace FI_ADDR_PROTO with FI_ADDR_UNSPEC
Allow apps to indicate that they don't care what address
format is used, and that the provider should indicate
the actual format.  Provider specific formats are usable
by setting the upper bit in the address format.  This
eliminates the need for FI_ADDR_PROTO.  Apps can still
treat the address as opaque if they care, but now the
real address format is provided back by the provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>