build: fix build for enable direct
Only one provider can be built in direct mode, so we
keep a count. Also, additional headers need to be
installed.

Include a check to see if the requested provider supports
direct mode, if the user requests to build the provider
as direct.

Quote message in AC_MSG_WARN in square brackets.

Fixes #194

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>