common: Provide ring buffer implementation
Add a simple ring buffer to the common code.  The ring buffer
requires serialization by the caller and that the caller check
for available data and handle any overruns.  The ring buffer
simply reads and writes data without checking that there is any
data to read or ensure that the write won't overwrite unread
data.

The implemetation does support parallel read and write
operations.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>