sockets: must include header files in tarball
The sockets provider .h files must (eventually) be listed in a
_SOURCES or EXTRA_DIST Makefile.am value so that Automake will
include them in distribution tarballs.

This commit removes the sym link for indexer.h, because there was no
-I set to prov/src/sockets to pick up that file, so the build was just
picking up include/indexer.h, anyway.

This raises another point: should indexer.c be sym linked to
prov/src/sockets?  It seems like this is another facett of the ongoing
"common sources" debate on #207.

This is part of the fix for #219.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>