fabric: Add mode bits to RX/TX attributes
When an RX/TX context is opened, an application can
specify a subset of the capabilities requested of the
endpoint.  As a result, the provider may be able to
provide relaxed mode bits.

For example, an endpoint that supports both tagged
messages and RMA operations may require FI_CONTEXT
mode bits.  However, a TX context that will only be
used for RMA may not require this mode.  By removing
this mode, an application may see improved performance
when using the RMA TX without needing to allocate
fi_context structures.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>