man: Put Markdown man pages in master
These man pages are now the Canonical Source of Truth.  Changes to the
man pages should be done in these Markdown man pages.

The Markdown man pages in the `gh-pages` branch are now copies of
these files, and are *not* definitive (i.e., changes to those files
may well be lost).