fabric: Remove EP sync call
Support for EP sync may be difficult for providers to implement,
as it requires providers to track all operations that have been
submitted by an app prior to the sync call being made.  For calls
that do not generate completion entries, there's no clear way for
the provider to know when the call has completed.

The use of a counter doesn't even help, as the provider must be
able to distinguish between calls issued before sync was
invoked, versus calls issued after sync is called, which is
difficult to track when completion may occur out of order.
Out of order completions is common for RDM endpoints and networks
that make use of adaptive routing algorithms.

The problems become more pronounced if the app does something
deranged like call sync after every request.

So, we basically need to punt this up to the application, which
can do a much better job of tracking which requests need to be
completed and has other optimization techniques available, such
as directing completions to different counters.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>