fabric: Remove provider free/dup info calls
The framework itself should provide the implementation for
fi_freeinfo and fi_dupinfo.  The API does not require that
an app use an fi_info structure that has been returned from
fi_getinfo in any of the function calls.  This restricts a
provider from attempting to do anything fancy in the
fi_getinfo call.

We can simplify the provider interface by only allowing
the framework to implement the free/dup info calls.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>