fabric: Merge connected and unconnected transfer calls
In an effort to reduce the number of interfaces exposed by
a provider, merge the connected and unconnected APIs.  The
result is instead of having separate calls, such as:

send, sendv, senddata, inject, sendmsg
sendto sendvto, senddatato, injectto, sendmsg

we have a single set:

send, sendv, senddata, inject, sendmsg

The actual interfaces are based on the unconnected calls.
The disadvantage is that connected endpoints will have an
extra parameter, which will be unused if scalable endpoints
are not in use.  The advantage is a smaller footprint for
supported APIs, with future support for scalable endpoints
over connected endpoints.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>