Static constructors for built-in providers v2
Add config.h defines for each built provider and if that provider is built as a DSO.

Add prov.h: a header which needs to be included near a provider's init function. It
defines the name of the init function, the init function's def, and symbols to call
the provider's ctor/dtor.

Signed-off-by: pmmccorm <patrick.m.mccormick@intel.com>