prov/psm: don't try name resolution if FI_SOURCE is set
The "node" parameter to fi_getinfo is usually the name/address of
the remote side. The provider can try to resolve it into a PSM
"epid" that can be used as the input to fi_av_insert.  However,
if the FI_SOURCE flags is set, the "node" parameter is the local
address and there is no need to try the name resolution (and in
fact it can't because the epid is only availabe after the domain
is open).

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>