fabric: create tx/rx ctx using scalable endpoints
This fixes some leftover issues in PR #340 that was
merged in c5a08e9.

The application creates a scalable endpoint, and must
use fid_sep in order to create the TX/RX contexts.

Also, include a separate fi_scalable_ep_bind() that
only documents binding an address vector to a scalable
endpoint.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>