fabric: Clarify which buffers FI_OPT_MIN_MULTI_RECV apply to
Improve documentation regarding the expected behavior when an
app modifies the MIN_MULTI_RECV endpoint option.

Recommend that the option be set before an endpoint is enabled,
in order to support the broadest array of implementations, and
that it only applies to buffers posted after the value has been
set.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>