prov/psm: remove FI_EP_MSG from supported ep types
Originally there was a type FID_MSG for reliable message based
transport, which was similar to RDM and was marked as one of the
supported types of the PSM provider. Later the transport types
were redefined. The new type FI_EP_MSG, although named similarly,
has totally different meaning (it means RC).

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>