fabric: Remove FI_EXTDIR environment variable
Remove all environment variables from the base libfabric
framework prior to the 1.0 release.  The FI_EXTDIR variable
is used to override where to search for external libaries,
which defaults to pkglibdir.

There is no pressing requirement for this variable, and it can
be added later if a need arises, once a proper framework
can be defined for handling environment variables.

Rename EXTDIR to better indicate how the directory is being
used.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>