prov/psm: remove an obsolete compatibility patch
The old patch provided backward compactibility of the fi_alais call
when FI_EVENT was changed from an EP flag to a CQ binding flag. Now
the patch not only is no longer needed, but also prevents FI_EVENT
being set as part of the default op_flags via the fi_alias call.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>