Fix RPM spec version number generated by nightly cron
This is a fix for issue #469, in which the generated tarball includes a
spec file with a version field containing dashes, which are disallowed.

This commit modifies the nightly cron job, so that it replaces the
.spec.in file with one which contains the correct version number.
Incidentally, it also removes the only substitution that is made by the
configure script, but since normal (non-daily) builds will not need this
modification, this redundancy should be ok.