sockets: fix check-the-hints logic
Check the hints for the specific EP type, not necessarily all of
them.  This prevents prematurely returning 0 if the EP hints are ok --
but the domain/fabric hints are not checked at all.

With the new logic, for example, all hints can be empty except
fabric->prov_name, and it will still be checked properly (i.e.,
sock_verify_fabric_attr()) will return -FI_ENODATA if the prov_name
does not match).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>