add fi_rx_size_left/fi_tx_size_left
This simply implements the -FI_ENOSYS version of these routines.  Follow
on commits will implement support in specific providers.

Reese's proposal and rationale for these operations is available in the
OFIWG mailing list archives:

http://lists.openfabrics.org/pipermail/ofiwg/2014-December/000477.html

Note that this implementation of the API uses "ssize_t" instead of
"size_t" return values.

Signed-off-by: Dave Goodell <dgoodell@cisco.com>