prov/psm: fix the handling of FI_MR_KEY and FI_MR_OFFSET flag
User defined MR key should only be used when FI_MR_KEY flag is
set and the FI_PROV_MR_ATTR mode is not set. Previously the
FI_MR_KEY flag was not checked.

This patch also corretcs an error in the address calculation for
MRs registered with the FI_MR_OFFSET flag.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>