Fix RPM spec version number generated by nightly cron
This fixes issue #469, in which the generated tarball includes a spec
file with a Version tag containing dashes, which are disallowed.

This commit modifies the nightly build script to replace the dashes in
the git describe output with periods, so that it can be used by RPM.
This affects both the tarball and RPM version numbers, but by changing
both, we can just let the spec file continue to use @VERSION@
substitution and avoid needing to maintain extra code to substitute a
transformed RPM version.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>