prov/psm: allow more than one endpoints to be RMA initiators
There can be only one PSM endpoint per process. The PSM provider
allows mutliple logical endpoints with the restriction that these
endpoints don't expose overlapping communication APIs. However,
for RMA and atomics, the restriction can be relaxed to allow
mutiple endpoints be RMA initiators. The new restriction for RMA
and atomics is that there can be only one endpoint to be used as
RMA target, and one endpoint to be used as atomics target. The
restriction on tagged and untagged message queue remain unchanged.

The related endpoint capabilities are defined as the following:

* FI_RMA only: endpoint can be used for RMA initiator and target
* FI_RMA plus any of the following flags:
	- FI_READ: endpoint can be used for RMA initiator
	- FI_WRITE: endpoint can be used for RMA initiator
	- FI_REMOTE_READ: endpoint can be used for RMA target
	- FI_REMOTE_WRITE: endpoint can be used for RMA target

* FI_ATOMICS only: endpoint can be atomics initiator and target
* FI_ATOMICS plus any of the following flags:
	- FI_READ: endpoint can be used for atomics initiator
	- FI_WRITE: endpoint can be used for atomics initiator
	- FI_REMOTE_READ: endpoint can be used for atomics target
	- FI_REMOTE_WRITE: endpoint can be used for atomics target

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>