prov/socket: Save domain service as char * instead of int
The domain service id is converted from a char * into an int.
But it is later used as a char * when binding the EP to the
domain.  This results in a failure when attempting to resolve
the domain's service ID.

Convert the domain service to a char *, which is consistent
with its usage and matches other fid's that also store a
service.  Remove (char *) casts around accessing the
service fields, so that the compiler can flag this error.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>