fabric: Replace fid_sep with fid_ep
The scalable endpoint fabric descriptor was changed from
fid_ep to fid_sep to clarify how it could be used.  However,
there is a desire to mix both scalable endpoints with
shared endpoints (for example, having multiple transmit
contexts, but using a shared receive context, and vice-versa).
This is almost possible, but if a scalable endpoint with
multiple receive contexts attempt to bind to a shared
transmit context, the result is that there is no FID onto
which to hang the send operations.

Neither fid_sep or fid_stx contains any data transfer ops.
The fix is to include data transfer ops as part of fid_sep,
which would make fid_sep identical to fid_ep.  Simply remove
the fid_sep structure and use fid_ep directly.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>