Fix empty md5sums.txt and sha1sums.txt
The doit function redirects all output to /dev/null when not in verbose
mode, which overrides the redirection of the md5sum/sha1sum output to
their respective files, so the files wind up empty.  This would not be
detectable if the script was tested in verbose mode, since the
redirections to /dev/null are not added in verbose mode.

This patch fixes the issue by adding an optional parameter to the doit
function which specifies the file to redirect standard output to.  If
the optional parameter is not specified and we are in verbose mode, no
redirection is done; otherwise, stdout is redirected to /dev/null.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>