fabric: Add FI_DISCARD search option
If a provider has buffered a received message, it may
be located using tsearch.  If the data for the buffered
message is no longer needed or out of date, the app
must currently search for the message, receive it into
a buffer, then discard the buffer.  It would be better
if the tsearch could simply mark that the message is
no longer needed, so that the provider can reclaim
the buffer.

This effectively allows an app to 'cancel' a buffered
receive.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>