add .mailmap file to fix git names+addresses
For a variety of reasons, several author and committer messages were
incorrect.  This includes:

- typos
- incorrect git client configuration
- GitHub unconditionally using your account's primary email address for
  web-based merge commits

The .mailmap file fixes the rendering in many git clients and in
git-shortlog output, including the default view in many web clients for
tags.

The difference with/without the .mailmap file can be seen by the
difference between the following commands:

```
{ git log --all --pretty=tformat:'%an <%ae>' ; \
  git log --all --pretty=tformat:'%cn <%ce>' } \
| sort | uniq
```
(shows without .mailmap)

and

```
{ git log --all --pretty=tformat:'%aN <%aE>' ; \
  git log --all --pretty=tformat:'%cN <%cE>' } \
| sort | uniq | wc -l
```
(shows with .mailmap)

Signed-off-by: Dave Goodell <dgoodell@cisco.com>