prov/verbs: Allow selective send completions
This commit adds support for the FI_COMPLETION flag for binding a
send completion queue to an endpoint and for selectively enabling
completions, and for the FI_COMPLETION flag in fi_tx_attr::op_flags,
which enables completions for all applicable transfer operations which
lack their own flags parameters.

These two settings are in separate fields in struct fi_ibv_msg_ep,
since they are independent---setting FI_COMPLETION in the flags
parameter to fi_ep_bind() has an effect on fi_sendmsg(), fi_writemsg(),
etc., but setting FI_COMPLETION in fi_tx_attr::op_flags does not affect
these operations!

In order to allow implementing fi_inject() and friends in the future,
this commit also disables the sq_sig_all flag when creating the queue
pair regardless of the value of FI_COMPLETION flag when binding the send
CQ.  This is because these operations never signal a completion, even if
the user does not set FI_COMPLETION when binding the send CQ.  Thus, we
must set the IBV_SEND_SIGNALED bit in all non-fi_inject*() work requests
if the FI_COMPLETION bit is not set in the ep_flags.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>