fabric/av: Remove support for inserting NULL address
A NULL address can be inserted into an AV in order to
indicate 'any' address (e.g. 0.0.0.0).  The intent was
that this could be used with recvfrom calls where the
buffer could be matched against a message from any
source.  However, FI_ADDR_UNSPEC was later added for
this reason.  This makes support for a NULL address
obsolete.  Simplify the AV API by removing support for
this as a documented requirement.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>