verbs: Fix fi_getinfo with src_addr but no dst_addr hint
The verbs provider would return -FI_EINVAL from fi_getinfo if node and
hints->dst_addr were NULL but src_addr was specified.  This occurred
because the verbs provider did not set the RAI_PASSIVE flag in its call
to rdma_getaddrinfo() in this case.  This commit fixes this issue by
setting the RAI_PASSIVE flag in this case.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>