cron-make-nightly-tarball.pl: Fix post-rc daily build versions
When this script was first written, the version in configure.ac and the
only tag in the repository were bogus and did not match each other.  Now
that we have an rc release properly tagged, versions generated by this
script will look like "1.0.0rc1.v1.0.0rc1.xyz.gabcdefab", which is a bit
redundant.  To fix this, use the git tag version as the main RPM version
instead of appending it to the version in configure.ac.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>