prov/sockets: Fix dependency on shm_open
Newer versions of glibc define clock_gettime in libc (it has
traditionally been in librt).  This means that librt is not
automatically pulled in by the check for clock_gettime, which causes
link errors for downstream projects trying to link against libfabric
under newer Linux versions (this issue was observed on Fedora 21).

This commit fixes the issue by adding an explicit check for shm_open in
the socket provider's configure-time checks.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>