fabric: ensure that all memory is freed if asprintf() fails
Rename "done" label to "libnl_done" to emphasize that it's in the

Add a loop at the libnl_done label to ensure that all liblist[]
entries are freed, and remove the additional free() when asprintf()
fails so that it's safe to jump down to libnl_done.  Also initialize
liblist and n at their declarations so that they are always
initialized -- e.g., even if dlopen().

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>