prov/verbs: Fix fcntl(fd, F_GETFL) usage
When creating an event queue or completion queue using an fd-based wait
object, we try to add the O_NONBLOCK flag to the set of flags, but we do
not check the return value of the first fcntl call, so we never actually
get the original fd flags (the third argument is ignored for F_GETFL)!
This commit fixes this issue so we actually retrieve the original flags
value.

This commit fixes Coverity IDs (CID) 29728 and 29739.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>